Moving generic functions to file misc.go
