+++ extractor/extractor.go
@@ -4,9 +4,7 @@ import (
 	"fmt"
 	"math/big"
 	"reflect"
-	"strings"
 
-	"github.com/mitchellh/copystructure"
 	"github.com/reilabs/gnark-lean-extractor/abstractor"
 
 	"github.com/consensys/gnark-crypto/ecc"
@@ -123,104 +121,6 @@ type ExGadget struct {
 
 func (g *ExGadget) isOp() {}
 
-func arrayToSlice(v reflect.Value) []frontend.Variable {
-	if v.Len() == 0 {
-		return []frontend.Variable{}
-	}
-
-	switch v.Index(0).Kind() {
-	case reflect.Array:
-		args := []frontend.Variable{}
-		for i := 0; i < v.Len(); i++ {
-			arg := arrayToSlice(v.Index(i))
-			// The reason to check for len != 0 is to avoid generating
-			// lists of empty nested lists
-			if len(arg) != 0 {
-				args = append(args, arg)
-			}
-		}
-		return args
-	case reflect.Interface:
-		res := []frontend.Variable{}
-		for i := 0; i < v.Len(); i++ {
-			res = append(res, v.Index(i).Elem().Interface().(frontend.Variable))
-		}
-		return res
-	default:
-		return []frontend.Variable{}
-	}
-}
-
-// flattenSlice takes a slice and returns a single dimension
-// slice of frontend.Variable. This is needed to transform
-// nested slices into single dimensional slices to be
-// processed by sanitizeVars.
-func flattenSlice(value reflect.Value) []frontend.Variable {
-	if value.Len() == 0 {
-		return []frontend.Variable{}
-	}
-	if value.Index(0).Kind() == reflect.Slice {
-		args := []frontend.Variable{}
-		for i := 0; i < value.Len(); i++ {
-			arg := flattenSlice(value.Index(i))
-			// The reason to check for len != 0 is to avoid generating
-			// lists of empty nested lists
-			if len(arg) != 0 {
-				args = append(args, arg)
-			}
-		}
-		return args
-	}
-	return value.Interface().([]frontend.Variable)
-}
-
-func updateProj(gate Operand, extra ...int) Proj {
-	if len(extra) == 2 {
-		return Proj{gate, extra[0], extra[1]}
-	}
-	return Proj{updateProj(gate, extra[:len(extra)-2]...), extra[len(extra)-2], extra[len(extra)-1]}
-}
-
-func replaceArg(gOutputs interface{}, gate Operand, extra ...int) interface{} {
-	// extra[0] -> i
-	// extra[1] -> len
-	switch v := (gOutputs).(type) {
-	case Input, Gate:
-		if len(extra) == 2 {
-			return Proj{gate, extra[0], extra[1]}
-		}
-		return gate
-	case Proj:
-		if len(extra) >= 2 {
-			return updateProj(gate, extra...)
-		}
-		return gate
-	case []frontend.Variable:
-		res := make([]frontend.Variable, len(v))
-		for i, o := range v {
-			res[i] = replaceArg(o, gate, append(extra, []int{i, len(v)}...)...)
-		}
-		return res
-	case [][]frontend.Variable:
-		res := make([][]frontend.Variable, len(v))
-		for i, o := range v {
-			res[i] = replaceArg(o, gate, append(extra, []int{i, len(v)}...)...).([]frontend.Variable)
-		}
-		return res
-	case [][][]frontend.Variable:
-		res := make([][][]frontend.Variable, len(v))
-		for i, o := range v {
-			res[i] = replaceArg(o, gate, append(extra, []int{i, len(v)}...)...).([][]frontend.Variable)
-		}
-		return res
-	case nil:
-		return []frontend.Variable{}
-	default:
-		fmt.Printf("invalid argument of type %T %#v\n", gOutputs, gOutputs)
-		panic("invalid argument")
-	}
-}
-
 func (g *ExGadget) Call(gadget abstractor.GadgetDefinition) interface{} {
 	args := []frontend.Variable{}
 
@@ -251,18 +151,6 @@ func (g *ExGadget) Call(gadget abstractor.GadgetDefinition) interface{} {
 	res := replaceArg(g.OutputsDeep, gate)
 	return res
 
-}
-
-func cloneGadget(gadget abstractor.GadgetDefinition) abstractor.GadgetDefinition {
-	dup, err := copystructure.Copy(gadget)
-	if err != nil {
-		panic(err)
-	}
-	// The reason for the following lines is to generate a reflect.Ptr to the interface
-	v := reflect.ValueOf(dup)
-	tmp_gadget := reflect.New(v.Type())
-	tmp_gadget.Elem().Set(v)
-	return tmp_gadget.Interface().(abstractor.GadgetDefinition)
 }
 
 func (ce *CodeExtractor) Call(gadget abstractor.GadgetDefinition) interface{} {
@@ -524,53 +412,6 @@ func (ce *CodeExtractor) ConstantValue(v frontend.Variable) (*big.Int, bool) {
 	default:
 		return nil, false
 	}
-}
-
-func generateUniqueName(element any, args []ExArg) string {
-	// To distinguish between gadgets instantiated with different array
-	// sizes, add a suffix to the name. The suffix of each instantiation
-	// is made up of the concatenation of the length of all the array
-	// fields in the gadget
-	suffix := ""
-	for _, a := range args {
-		if a.Kind == reflect.Array || a.Kind == reflect.Slice {
-			suffix += "_"
-			suffix += strings.Join(getSizeGadgetArgs(a.Type), "_")
-		}
-	}
-
-	val := reflect.ValueOf(element).Elem()
-	for i := 0; i < val.NumField(); i++ {
-		switch val.Field(i).Kind() {
-		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-			suffix += fmt.Sprintf("_%d", val.Field(i).Int())
-		case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-			suffix += fmt.Sprintf("_%d", val.Field(i).Uint())
-		case reflect.Uintptr, reflect.Float32, reflect.Float64, reflect.Complex64, reflect.Complex128:
-			fmt.Printf("-- Gadget name doesn't differentiate yet between different initialised values of type %+v.\n", val.Field(i).Kind())
-			fmt.Println("-- Proceed with caution")
-		}
-	}
-	return fmt.Sprintf("%s%s", reflect.TypeOf(element).Elem().Name(), suffix)
-}
-
-func getGadgetByName(gadgets []ExGadget, name string) abstractor.Gadget {
-	for _, gadget := range gadgets {
-		if gadget.Name == name {
-			return &gadget
-		}
-	}
-	return nil
-}
-
-// getSizeGadgetArgs generates the concatenation of dimensions of
-// a slice/array (i.e. [3][2]frontend.Variable --> ["3","2"])
-// It is used to generate a unique gadget name
-func getSizeGadgetArgs(elem ExArgType) []string {
-	if elem.Type == nil {
-		return []string{fmt.Sprintf("%d", elem.Size)}
-	}
-	return append(getSizeGadgetArgs(*elem.Type), fmt.Sprintf("%d", elem.Size))
 }
 
 func (ce *CodeExtractor) DefineGadget(gadget abstractor.GadgetDefinition) abstractor.Gadget {
